Revert "Merge branch 'Feature-BuilderGUI-(TrafficLights)' into Feature-TrafficLights"
This reverts commit 23ba7b8cd9d6556e072662d3b7c7a0517a3f6e95.
